refactor: Move node params resolver to another file
